Output poses to 3d view

+++ src/Odometry/Matrix.h
@@ -43,8 +43,8 @@ Street, Fifth Floor, Boston, MA 02110-1301, USA
 
 #define endll endl << endl // double end line definition
 
-typedef double FLOAT;      // double precision
-//typedef float  FLOAT;    // single precision
+//typedef double FLOAT;      // double precision
+typedef float  FLOAT;    // single precision
 
 class Matrix {
